Make the middle of the center key post thinner
There tends to be some layer irregularities on the post where the magnet
hole ends. With a straight post, those irregularies have to be filed/cut
down for a proper fit. This reduces the impact of the irregularities,
so that only the front face of the post (with the magnet) should need to
be cleaned up.