Remove some of the extraneous exported files
This removes the exported assembly files and the stls for the pcb objects.
These aren't really needed, and take up quite a bit of space.

The assembly files in particular are the largest in the repository, and
the f3ds were being needlessly updated after every export.

The assemblies are still being exported to the fusion cloud though, for
convenient viewing on the web.